+++ js/bloodsplats.js
@@ -83,23 +83,25 @@ export class BloodSplats {
             for (let effect of this.hud.effects.children) {
                 effect.alpha = 0;
             }
-            if (['dnd5e.LootSheet5eNPC', 'core.a'].includes(this.actor?.getFlag("core", "sheetClass")) || this.actor?.getFlag("item-piles", "data.enabled") == true) {
+            if (['dnd5e.LootSheetNPC5e', 'core.MerchantSheet'].includes(this.actor?.data?.flags?.core?.sheetClass) || this.actor?.data.flags["item-piles"]?.data?.enabled == true) {
                 this.icon.alpha = 0.5;
                 if (this.bloodsplat) {
                     this.removeChild(this.bloodsplat);
                     delete this.bloodsplat;
                 }
-                if (this.actor?.getFlag("item-piles", "data.enabled") !== true) {
+                if (this.actor?.data.flags["item-piles"]?.data?.enabled !== true) {
                     if (this.tresurechest == undefined) {
-                        loadTexture("icons/svg/chest.svg").then((tex) => { //"modules/monks-little-details/img/chest.png"
-                            const chesticon = new PIXI.Sprite(tex);
-                            const size = Math.min(canvas.grid.grid.w, canvas.grid.grid.h);
-                            chesticon.width = chesticon.height = size;
-                            chesticon.position.set((this.w - size) / 2, (this.h - size) / 2);
-                            chesticon.alpha = 0.8;
-                            this.tresurechest = chesticon;
-                            this.addChild(this.tresurechest);
-                        });
+                        if (setting("treasure-chest") != "") {
+                            loadTexture(setting("treasure-chest")).then((tex) => {
+                                const chesticon = new PIXI.Sprite(tex);
+                                const size = Math.min(canvas.grid.grid.w, canvas.grid.grid.h);
+                                chesticon.width = chesticon.height = size * setting("treasure-chest-size");
+                                chesticon.position.set((this.w - chesticon.width) / 2, (this.h - chesticon.height) / 2);
+                                chesticon.alpha = 0.8;
+                                this.tresurechest = chesticon;
+                                this.addChild(this.tresurechest);
+                            });
+                        }
                     } else
                         this.tresurechest.alpha = (this._hover ? 1 : 0.8);
                 } else {